{{cookiecutter.project_name}}/requirements.txt
aiofiles==0.5.0
aioredis==1.3.1
alembic==1.4.2
async-timeout==3.0.1
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
ecdsa==0.15
email-validator==1.1.1
fastapi==0.61.1
gunicorn==20.0.4
h11==0.9.0
hiredis==1.1.0
httptools==0.1.1
idna==2.9
loguru==0.5.1
Mako==1.1.3
MarkupSafe==1.1.1
passlib==1.7.2
pyasn1==0.4.8
pycparser==2.20
pydantic==1.5.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
python-multipart==0.0.5
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.17
starlette==0.13.6
uvicorn==0.11.7
uvloop==0.14.0
websockets==8.1